pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "cognos-enterprise"
version = "1.0.0"
description = "CognOS Enterprise Gateway — Pluggable LLM trust-scoring with multi-tenant isolation"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
pdf = ["reportlab>=4.0.0"]
oauth = ["python-jose[cryptography]>=3.3.0", "authlib>=1.3.0"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "httpx>=0.27.0"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
